+++ apps/web/src/components/BaseLayout.vue
@@ -21,7 +21,7 @@
 </template>
 
 <script setup lang="ts">
-import { useTheme } from "@/composables/useTheme";
+import { useTheme } from "@/composables";
 
 const { isLightTheme } = useTheme();
 </script>
