+++ apps/web/src/components/ToggleTheme.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup lang="ts">
-import { useTheme } from "@/composables/useTheme";
+import { useTheme } from "@/composables";
 import SwitchUI from "./ui/SwitchUI.vue";
 
 const { toggleTheme, isLightTheme } = useTheme();
